requirements.txt
numpy
scipy
pandas
scikit-learn
streamlit
plotly
Pillow
shap
matplotlib
lightgbm